# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Categories Controller",
        "version": "1"
    },
    "paths": {
...
        "/admin/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "operationId": "d3b2e17f118203ba1ae8955cbebb733b",
                "responses": {
                    "200": {
                        "description": "product fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminProduct"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "products"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "541884d6a9ac68d92989c2e4a328900d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/adminUpdateProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product was updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminProduct"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminProductIdPathParameter"
                }
            ]
        },
        "/admin/products/{id}/activate": {
            "patch": {
                "tags": [
                    "products"
                ],
                "operationId": "06a67b6cb55a8221505a0652b2557050",
                "responses": {
                    "204": {
                        "description": "Fetched Product Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminProductList"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminProductIdPathParameter"
                }
            ]
        },
        "/admin/products/{id}/deActivate": {
            "patch": {
                "tags": [
                    "products"
                ],
                "operationId": "585976e57474020fea56e3255fa088d8",
                "responses": {
                    "204": {
                        "description": "Fetched Product Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminProductList"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminProductIdPathParameter"
                }
            ]
        },
...
        "/admin/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "operationId": "20ecf04d8244289b7896b1a6e7c380d1",
                "responses": {
                    "200": {
                        "description": "The Product was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminProduct"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "operationId": "3fc4718ebf0611de6131a6d4da005058",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/adminCreateProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Fetched Product Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminProductList"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/products/getProductsByName": {
            "post": {
                "tags": [
                    "products"
                ],
                "operationId": "fb1c59ca555c3bb16d18e6f2ab848fce",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/adminProductName"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Fetched Product Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminProductList"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "adminCategory": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "parent_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "default": "2017-02-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/adminCategory"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "adminCreateProduct": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string",
                        "example": "00.00"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_most_buy": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "unit": {
                        "$ref": "#/components/schemas/Unit"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "adminProduct": {
                "properties": {
                    "id": {
                        "description": "CategoryData>",
                        "type": "integer"
                    },
                    "name": {
                        "description": "CategoryData>",
                        "type": "string"
                    },
                    "price": {
                        "description": "CategoryData>",
                        "type": "string"
                    },
                    "description": {
                        "description": "CategoryData>",
                        "type": "string"
                    },
                    "price_offer": {
                        "description": "CategoryData>",
                        "type": "string",
                        "nullable": true
                    },
                    "is_most_buy": {
                        "description": "CategoryData>",
                        "type": "boolean"
                    },
                    "is_active": {
                        "description": "CategoryData>",
                        "type": "boolean"
                    },
                    "unit": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Unit"
                            }
                        ],
                        "nullable": true
                    },
                    "unit_value": {
                        "description": "CategoryData>",
                        "type": "integer",
                        "nullable": true
                    },
                    "image": {
                        "description": "CategoryData>",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "CategoryData>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2017-02-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "childCategories": {
                        "description": "CategoryData>",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/adminCategory"
                        }
                    },
                    "parentCategories": {
                        "description": "CategoryData>",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/adminCategory"
                        }
                    }
                },
                "type": "object"
            },
            "adminProductList": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "adminUpdateProduct": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string",
                        "example": "00.00"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_most_buy": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "unit": {
                        "$ref": "#/components/schemas/Unit"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Unit": {
                "description": "[1 => Kg, 2 => K, 3 => L, 4 => Ml, 5 => Stock]",
                "type": "integer",
                "enum": [
                    1,
                    2,
                    3,
                    4,
                    5
                ]
            }
        },
        "parameters": {
...
            "adminProductIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "adminProductName": {
                "name": "name",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ]
}